Build base docker image in CD workflow
Add CD workflow to run on push to main. It builds and pushes the
base image to registry. This centralized image is used in other
dockerfiles as base image.
Change electron to use the docker-compose.yml file rather than
the docker-compose-prod. We need to add registry for the other images
used. We also need to find a way to use prod compose file in the
executable electron app, not in dev.